Refactor object validation
