+++ src/RPKI/Meta/GitVersionInfo.hs
@@ -5,13 +5,13 @@ gitBranch :: String
 gitBranch = "fix-cache-poisoning"
 
 gitHash :: String
-gitHash = "43cfe8573a2b7062646386e0295706fe2fbe2b3c"
+gitHash = "a1043a2b38d2f50169b4c2ea59f12edbb3a90789"
 
 gitCommitDate :: String
-gitCommitDate = "2026-08-10 21:06:46 +0200"
+gitCommitDate = "2026-08-10 21:07:44 +0200"
 
 gitCommitCount :: Int
-gitCommitCount = 2695
+gitCommitCount = 2696
 
 gitDirty :: Bool
 gitDirty = True
