+++ src/RPKI/Meta/UniqueId.hs
@@ -9,5 +9,5 @@ thisExecutableVersion :: ExecutableVersion
 thisExecutableVersion = ExecutableVersion $ rpkiProverVersion <> " " <> 
     -- The content is to be updated by the 'src-hash' script 
     -- that calculates hash of the source tree and configuration/build files     
-    "srcHash#e48fec5b782346a154fdba87ecb962a6820d4ab7fadc8b609168d36da0916bdf#srcHash"
+    "srcHash#bf97576591cdbb430ba5d1bacae80f2516b0a3673d59c49bff61768b2e0d2570#srcHash"
