+++ src/RPKI/Rsync.hs
@@ -46,7 +46,6 @@ import           RPKI.Store.Base.Storage
 import qualified RPKI.Store.Database    as DB
 import           RPKI.Time
 import qualified RPKI.Util                        as U
-import           RPKI.Validation.Common  (prevalidateObject)
 import           RPKI.Validation.ObjectValidation
 import           RPKI.Worker
